requirements.txt
gunicorn~=20.1.0
python-dotenv~=0.20.0
Flask~=2.1.2
Flask-SQLAlchemy~=2.5.1
SQLAlchemy~=1.4.37
psycopg2-binary~=2.9.3 ; sys.platform != 'linux'
psycopg2~=2.9.3 ; sys.platform == 'linux'
